Add AutoResetEvent and pull MessageBoxViewModel out into its own file
